Allow ssh git links to be used.
Links in the form of git@github.com:lsegal/yard.git
are now the default on github so it makes sense to
have the example in that format too.

Supporting SSH links makes running a private instance
nicer.

Mostly a duplication of #71 but drops the config
switch stuff in favour of just using the ssh format
everywhere.