Remove pagy due to removed support for Rails in v43.x
pagy v43 removed out of the box support for standard Rails param routing which means we must now use workarounds to implement proper pagination with path based routes. This as seen as an unnecessary maintenance burden and isn't recommended by the project maintainer. Given the uncertainty of future support for Rails, it's not worth continuing to use pagy. See ddnexus/pagy#846

The good news is that pagy is easily replaceable and does not seem to be doing very much for our use case, so we can just use out of the box Rails methods.